[BE] REFACTOR: calDateDiff 메소드 호출 경로 변경 #624
