+++ backend/src/auth/auth.controller.ts
@@ -69,7 +69,7 @@ export class AuthController {
     // NOTE: 42 계정이 존재하면 무조건 로그인 처리를 할것이므로 계정 등록도 여기서 처리합니다.
     await this.authService.addUserIfNotExists(user);
     return res.redirect(
-      `https://${this.configService.get<string>(
+      `http://${this.configService.get<string>(
         'fe_host',
       )}/main?access_token=${token}`,
     );
