+++ backend/src/utils/utils.module.ts
@@ -1,8 +1,7 @@
 import { MailerModule } from '@nestjs-modules/mailer';
-import { Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { ScheduleModule } from '@nestjs/schedule';
-import { BanModule } from 'src/ban/ban.module';
 import MailerConfigService from 'src/config/mailer.config';
 import { LentModule } from 'src/lent/lent.module';
 import { CabinetModule } from '../cabinet/cabinet.module';
@@ -18,8 +17,7 @@ import { ExpiredChecker } from './expired.checker.component';
       useClass: MailerConfigService,
     }),
     CabinetModule,
-    LentModule,
-    BanModule,
+    forwardRef(() => LentModule),
   ],
 
   providers: [EmailSender, ExpiredChecker, DateCalculator],
